feat: aap-38589 move to github_app.py
