+++ src/awx_plugins/credentials/plugins.py
@@ -7,7 +7,7 @@ from awx_plugins.interfaces._temporary_private_api import (  # noqa: WPS436
 from awx_plugins.interfaces._temporary_private_django_api import (  # noqa: WPS436
     gettext_noop,
 )
-from github import Github, Auth
+# from github import Github, Auth
 from urllib.parse import urlparse
 
 from .injectors import (
@@ -537,118 +537,118 @@ github_token = ManagedCredentialType(
     },
 )
 
-github_app_inputs={
-    'fields': [
-    {
-        'id': 'github_url',
-        'label': 'GitHub URL',
-        'type': 'string',
-        'help_text': gettext_noop(
-            'Specify the GitHub URL here in the case of an Enterprise Github '
-            'Example: https://example.github.com',
-        ),
-        'default': 'https://github.com',
-    },
-    {
-        'id': 'app_id',
-        'label': 'Github App id created by the GitHub Admin',
-        'type': 'string',
-        'help_text': gettext_noop(
-            'The GitHub App Id created by the GitHub Admin '
-            'Example App ID: 1121547 '
-            'from https://github.com/settings/apps/arestlel-github-app-ansible'
-            'required for creating a bearer token for authentication',
-        ),
-        'default': '1121547',
-    },
-    {
-        'id': 'install_id',
-        'label': 'Installation id',
-        'type': 'string',
-        'help_text': gettext_noop(
-            'The  Installation ID from the GitHub App installation generated by the GitHub Admin '
-            'Example: 59980338 '
-            'extracted from the installation link: https://github.com/settings/installations/59980338 '
-            'required for creating a limited github app token'
-        ),
-        'default': '59980338',
-    },                      
-    {
-        'id': 'ssh_key_data',
-        'label': gettext_noop('RSA Private Key'),
-        'type': 'string',
-        'format': 'ssh_private_key',
-        'secret': True,
-        'multiline': True,
-        'help_text': gettext_noop(
-            'Paste the contents of the PEM file that the GitHub Admin '
-            'provided to you with the app and installation ids.',
-        ),
-    },
-    ],
-    'required': ['app_id','install_id','ssh_key_data',],
-    }
+# github_app_inputs={
+#     'fields': [
+#     {
+#         'id': 'github_url',
+#         'label': 'GitHub URL',
+#         'type': 'string',
+#         'help_text': gettext_noop(
+#             'Specify the GitHub URL here in the case of an Enterprise Github '
+#             'Example: https://example.github.com',
+#         ),
+#         'default': 'https://github.com',
+#     },
+#     {
+#         'id': 'app_id',
+#         'label': 'Github App id created by the GitHub Admin',
+#         'type': 'string',
+#         'help_text': gettext_noop(
+#             'The GitHub App Id created by the GitHub Admin '
+#             'Example App ID: 1121547 '
+#             'from https://github.com/settings/apps/arestlel-github-app-ansible'
+#             'required for creating a bearer token for authentication',
+#         ),
+#         'default': '1121547',
+#     },
+#     {
+#         'id': 'install_id',
+#         'label': 'Installation id',
+#         'type': 'string',
+#         'help_text': gettext_noop(
+#             'The  Installation ID from the GitHub App installation generated by the GitHub Admin '
+#             'Example: 59980338 '
+#             'extracted from the installation link: https://github.com/settings/installations/59980338 '
+#             'required for creating a limited github app token'
+#         ),
+#         'default': '59980338',
+#     },                      
+#     {
+#         'id': 'ssh_key_data',
+#         'label': gettext_noop('RSA Private Key'),
+#         'type': 'string',
+#         'format': 'ssh_private_key',
+#         'secret': True,
+#         'multiline': True,
+#         'help_text': gettext_noop(
+#             'Paste the contents of the PEM file that the GitHub Admin '
+#             'provided to you with the app and installation ids.',
+#         ),
+#     },
+#     ],
+#     'required': ['app_id','install_id','ssh_key_data',],
+#     }
 
 
-github_app_token = ManagedCredentialType(
-    namespace='github_app_token',
-    kind='scm', # scm source control management so it can be used for github authentication.
-    name=gettext_noop('GitHub App Token'),
-    managed=True,
-    inputs=github_app_inputs,
-    custom_injectors=github_app_injector,
-)
+# github_app_token = ManagedCredentialType(
+#     namespace='github_app_token',
+#     kind='scm', # scm source control management so it can be used for github authentication.
+#     name=gettext_noop('GitHub App Token'),
+#     managed=True,
+#     inputs=github_app_inputs,
+#     custom_injectors=github_app_injector,
+# )
 
-def github_app_backend(**kwargs):
-    github_url = kwargs.get("github_url")
-    app_id = kwargs.get("app_id")
-    installation_id = kwargs.get("install_id")
-    private_key = kwargs.get("ssh_key_data")
-    jwt_expiry = kwargs.get("jwt_expiry", 600)
+# def github_app_backend(**kwargs):
+#     github_url = kwargs.get("github_url")
+#     app_id = kwargs.get("app_id")
+#     installation_id = kwargs.get("install_id")
+#     private_key = kwargs.get("ssh_key_data")
+#     jwt_expiry = kwargs.get("jwt_expiry", 600)
 
-    missing_parameters = []
-    if not github_url:
-        missing_parameters.append("GitHub URL")
-    if not installation_id:
-        missing_parameters.append("Installation ID")
-    if not app_id:
-        missing_parameters.append("Application ID")
-    if not private_key:
-        missing_parameters.append("Private Key")
-    if missing_parameters:
-        raise Exception (f"Missing Parameter: {missing_parameters}")
+#     missing_parameters = []
+#     if not github_url:
+#         missing_parameters.append("GitHub URL")
+#     if not installation_id:
+#         missing_parameters.append("Installation ID")
+#     if not app_id:
+#         missing_parameters.append("Application ID")
+#     if not private_key:
+#         missing_parameters.append("Private Key")
+#     if missing_parameters:
+#         raise Exception (f"Missing Parameter: {missing_parameters}")
     
-    # verify the installation id andn the app id are integers
-    try:
-        app_id = int(app_id)
-        installation_id = int(installation_id)
-    except ValueError:
-        raise Exception("App ID and Installation ID must be integers")
+#     # verify the installation id andn the app id are integers
+#     try:
+#         app_id = int(app_id)
+#         installation_id = int(installation_id)
+#     except ValueError:
+#         raise Exception("App ID and Installation ID must be integers")
     
-    try:
-        jwt_expiry = int(jwt_expiry)
-    except ValueError:
-        raise Exception("JWT Expiry must be an integer")
+#     try:
+#         jwt_expiry = int(jwt_expiry)
+#     except ValueError:
+#         raise Exception("JWT Expiry must be an integer")
     
-    try:
-        github_url = str(github_url)
-        # make sure it parses as a URL
-        parsed_url = urlparse(github_url)
-    except ValueError:
-        raise Exception.InvalidParameterError("GitHub URL must be a string")
+#     try:
+#         github_url = str(github_url)
+#         # make sure it parses as a URL
+#         parsed_url = urlparse(github_url)
+#     except ValueError:
+#         raise Exception.InvalidParameterError("GitHub URL must be a string")
 
-    auth = Auth.AppAuth(app_id, private_key).get_installation_auth(installation_id, None)
+#     auth = Auth.AppAuth(app_id, private_key).get_installation_auth(installation_id, None)
 
-    # In order to generate a token we have to make an initial call
-    # to GitHub even though we dont need to interact with GitHub.
-    Github(auth=auth)
+#     # In order to generate a token we have to make an initial call
+#     # to GitHub even though we dont need to interact with GitHub.
+#     Github(auth=auth)
 
-    return auth.token
+#     return auth.token
 
-github_app_lookup = CredentialPlugin(
-    "GitHub App Lookup", 
-    inputs=github_app_inputs, 
-    backend=github_app_backend)
+# github_app_lookup = CredentialPlugin(
+#     "GitHub App Lookup", 
+#     inputs=github_app_inputs, 
+#     backend=github_app_backend)
 
 gitlab_token = ManagedCredentialType(
     namespace='gitlab_token',
